rnaseq-app/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.34
```txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1